Add an option to force Mir off
This will be used to handle the special case of hybrid
graphics with fglrx and nvidia, in which a Mir compatible
driver (e.g. intel) is used as an output sink to display
what was rendered by another display device which, in
turn, uses a Mir incompatible driver (e.g. fglrx or
nvidia).